databricks/tfl-analytics-project/requirements.txt
azure-core==1.39.0
azure-storage-blob==12.28.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
cryptography==46.0.7
idna==3.11
isodate==0.7.2
numpy==2.0.2
packaging==26.0
pandas==2.3.3
pycparser==2.23
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2026.1.post1
requests==2.32.5
six==1.17.0
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
